refactor(interp): changes to the os spefic handling of commands
* move from `handle_{os}.go` to `os_unix.go` and `os_notunix.go`
* use `sys/unix` instead of `syscall`
* share code between windows and plan 9
